% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_date.R
\name{add_date}
\alias{add_date}
\title{add_date}
\usage{
add_date(f, d)
}
\arguments{
\item{f}{Array with frequencies}

\item{d}{Array of dates}
}
\description{
Adds dates in wich the weather types are obtained
}
\keyword{internal}
