% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_constant.R
\name{calculate_constant}
\alias{calculate_constant}
\title{calculate_constant}
\usage{
calculate_constant(lon, lat)
}
\arguments{
\item{lon}{Numeric longitude of the central point}

\item{lat}{Numeric latitude of the central point}
}
\description{
Calculates constants required to apply the rules
}
\keyword{internal}
