% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_frequency.R
\name{calculate_frequency}
\alias{calculate_frequency}
\title{calculate_frequency}
\usage{
calculate_frequency(datalist, purelist, let)
}
\arguments{
\item{datalist}{List with the hybrid types}

\item{purelist}{List with the rest of types}

\item{let}{String with the name of the type}
}
\description{
Computes the frequency of weather types, after dividing the
             frequency of the hybrid types
}
\keyword{internal}
