% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_div.R
\name{get_div}
\alias{get_div}
\title{get_div}
\usage{
get_div(fr, nam)
}
\arguments{
\item{fr}{Numeric with the frequency values}

\item{nam}{String with the name of the weather type}
}
\description{
Divides the frequency of the hybrid type according to the rules # TODO: which rules?
}
\keyword{internal}
