% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jcpoints.R
\name{get_jcpoints}
\alias{get_jcpoints}
\title{get_jcpoints}
\usage{
get_jcpoints(lon, lat)
}
\arguments{
\item{lon}{Longitude of the central point.}

\item{lat}{Latitude of the central point.}
}
\value{
A data frame with the coordinates of the 16-grid points and the central point.
}
\description{
Computes the 16-grid points that determine the JC-scheme, based on the
             coordinates of the central point.
}
\details{
The function excludes the poles and the equatorial areas between 25S-25N.
}
\seealso{
\code{\link{plot_jcscheme}}
}
