% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmat_cwt.R
\name{getmat_cwt}
\alias{getmat_cwt}
\title{getmat_cwt}
\usage{
getmat_cwt(cwt_out, centralp, times, out = c("CWT", "indices"))
}
\arguments{
\item{cwt_out}{List with the frequencies of weather types and indices}

\item{centralp}{List with multiples central points}

\item{times}{Array with the dates}

\item{out}{String to select the output}
}
\description{
Puts in a matrix the resulting list
}
\keyword{internal}
