% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getname_hyb.R
\name{getname_hyb}
\alias{getname_hyb}
\title{getname_hyb}
\usage{
getname_hyb(hybrid_div)
}
\arguments{
\item{hybrid_div}{List with the hybrid types}
}
\description{
Gets the name of the hybrid types to account for the rest of the types
}
\keyword{internal}
