% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isnot_validcp.R
\name{isnot_validcp}
\alias{isnot_validcp}
\title{isnot_validcp}
\usage{
isnot_validcp(grid16, loni, lati)
}
\arguments{
\item{grid16}{Data.frame with the 16-grid scheme}

\item{loni}{Array with longitudes}

\item{lati}{Array with latitudes}
}
\description{
Checks if the central point used is valid.
}
\keyword{internal}
