% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ncdata.R
\name{read_ncdata}
\alias{read_ncdata}
\title{read_ncdata}
\usage{
read_ncdata(ncinput, nam_coor, units)
}
\arguments{
\item{ncinput}{A NETCDF file with pressure field.}

\item{nam_coor}{Names of space and time coordinates}

\item{units}{Units required (Pa or hPa).}
}
\value{
A list with:
\itemize{
\item A 3D-array of mean sea level (or pressure field) as [lon,lat,times]. The units returned as hPa.
\item A numeric with longitudes values.
\item A numeric with latitudes values.
\item A numeric with dates values.
}
}
\description{
Reads a ncdf input file to extract the input for the classification:
             pressure field, longitudes, latitudes and dates.
             Absolute time is required to read the dates properly.
}
